refactor: component/common 내부 폴더 구조 변경, LoginForm (타입이름 변경)
