+++ src/components/common/layout/CommonLayout.tsx
@@ -1,4 +1,4 @@
-import { Sidebar } from './components/Sidebar';
+import { Sidebar } from '../Sidebar';
 
 export const CommonLayout = () => {
   return (
