+++ src/pages/MainPage.tsx
@@ -1,4 +1,4 @@
-import { CommonLayout } from '@/components/common/CommonLayout';
+import { CommonLayout } from '@/components/common/layout/CommonLayout';
 
 export const MainPage = () => {
   return <CommonLayout />;
